chart_coordinator_project/requirements.txt
# ===================================================================
# Chart Coordinator Project - Final Dependencies for Deployment
# 项目最终部署依赖清单 (V1.0)
# - 该文件由AI助手通过代码扫描和人工分析生成
# - 包含了运行所有17个渲染工具所需的全部核心依赖
# - 更新日期: 2024-06-20
# ===================================================================

# --- 核心框架 (Core Framework) ---
google-genai>=0.7.0          # Google ADK/Gemini 核心
google-adk>=0.1.0            # Google ADK 框架
python-dotenv>=1.0.0       # 用于加载 .env 环境变量

# --- 基础数据科学库 (Essential Data Science) ---
numpy>=1.24.0              # 数值计算
pandas>=2.0.0              # 数据处理
scipy>=1.10.0              # 科学计算 (被多个库隐式依赖)

# --- 2D/3D 可视化 (2D/3D Visualization) ---
matplotlib>=3.7.0          # 2D绘图基础库
seaborn>=0.12.0            # 基于Matplotlib的统计可视化
plotly>=5.15.0             # 交互式图表
kaleido>=0.2.1             # Plotly 静态图像导出
mplfinance==0.12.10b0      # 金融K线图 (使用最新的beta版本)
py3Dmol>=2.0.0             # 3D分子结构可视化

# --- 网络图与流程图 (Network & Flowcharts) ---
networkx>=3.0              # 图论与网络分析
pyvis>=0.3.0               # 交互式网络图(基于vis.js)
graphviz>=0.20.0           # Graphviz渲染引擎的Python接口
# 注意: Graphviz 工具还需要在操作系统中安装 Graphviz 软件
# (e.g., via `choco install graphviz` on Windows or `brew install graphviz` on macOS)

# --- 地理空间可视化 (Geospatial Visualization) ---
folium>=0.19.0             # 交互式地图 (基于Leaflet.js)
# geopandas>=1.0.0         # (可选) 如果需要处理复杂的地理数据格式

# --- Web与JavaScript集成 (Web & JavaScript Integration) ---
selenium>=4.0.0            # 驱动浏览器进行截图等操作
chromedriver-autoinstaller>=0.6.0  # 自动下载匹配版本的ChromeDriver
lxml>=5.0.0                # 高性能的XML/HTML解析器
beautifulsoup4>=4.12.0     # 用于解析HTML，提取图表内容
# js2py>=0.74              # (可选) 如果需要直接在Python中运行JS片段

# --- 其它工具库 (Other Utilities) ---
requests>=2.31.0           # 用于HTTP请求 (例如: PlantUML服务器)
pillow>=10.0.0             # 图像处理库 (保存、转换格式)
statsmodels>=0.14.0        # 统计模型 (用于生成趋势线等)
typing-extensions>=4.5.0   # 兼容旧版本Python的类型提示

# ===================================================================
# 系统级依赖 (System-level Dependencies)
# ===================================================================
# 1. Node.js (LTS Version): 用于执行 ECharts, Three.js, Flowchart.js 等
#    - 安装后，在 `chart_coordinator_project` 目录下运行: npm install
#
# 2. Graphviz: 用于渲染 Graphviz 图表
#    - Windows: choco install graphviz
#    - macOS: brew install graphviz
#    - Linux (Ubuntu): sudo apt-get install graphviz
#
# 3. Java Runtime Environment (JRE 8+): 用于运行 PlantUML
#    - PlantUMLRenderTool 会从服务器请求渲染，但本地渲染需要Java环境
#
# 4. 中文字体 (Chinese Fonts): 如果图表需要显示中文
#    - 需要在操作系统中安装中文字体 (如 "SimHei", "Microsoft YaHei" 等)
#      并可能需要清理matplotlib缓存。
# =================================================================== 

litellm>=1.0.0                # ADK依赖的LiteLLM库 